requirements.txt
python-chess==1.999
numpy
torch --index-url https://download.pytorch.org/whl/cpu
